+++ visual-dotnet/SauceLabs.Visual/VisualClientBase.cs
@@ -2,6 +2,8 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Polly;
+using Polly.Retry;
 using SauceLabs.Visual.GraphQL;
 using SauceLabs.Visual.Models;
 using SauceLabs.Visual.Utils;
@@ -18,10 +20,26 @@ namespace SauceLabs.Visual
         public BaselineOverride? BaselineOverride { get; set; }
 
         protected string? PreviousSuiteName = null;
+        private readonly ResiliencePipeline _retryPipeline;
 
         internal VisualClientBase(Region region, string username, string accessKey)
         {
+            if (StringUtils.IsNullOrEmpty(username) || StringUtils.IsNullOrEmpty(accessKey))
+            {
+                throw new VisualClientException("Username or Access Key not set");
+            }
+
             Api = new VisualApi(region, username, accessKey);
+
+            _retryPipeline = new ResiliencePipelineBuilder()
+                .AddRetry(new RetryStrategyOptions()
+                {
+                    Name = "VisualRetryPolicy",
+                    Delay = TimeSpan.FromSeconds(1),
+                    MaxRetryAttempts = 10
+                })
+                .AddTimeout(TimeSpan.FromSeconds(15))
+                .Build();
         }
 
         internal async Task<string> VisualCheckBaseAsync(string name,
@@ -60,6 +78,45 @@ namespace SauceLabs.Visual
         public void Dispose()
         {
             Api.Dispose();
+        }
+
+
+        /// <summary>
+        /// <c>VisualResults</c> returns the results of screenshot comparison.
+        /// </summary>
+        /// <returns>a dictionary containing <c>DiffStatus</c> and the number of screenshot in that status.</returns>
+        /// <exception cref="VisualClientException"></exception>
+        public async Task<Dictionary<DiffStatus, int>> VisualResults()
+        {
+            return await _retryPipeline.ExecuteAsync(async token => await FetchVisualResults(Build.Id));
+        }
+
+        private async Task<Dictionary<DiffStatus, int>> FetchVisualResults(string buildId)
+        {
+            var dict = new Dictionary<DiffStatus, int>() {
+                { DiffStatus.Approved, 0 },
+                { DiffStatus.Equal, 0 },
+                { DiffStatus.Unapproved, 0 },
+                { DiffStatus.Errored, 0 },
+                { DiffStatus.Queued, 0 },
+                { DiffStatus.Rejected, 0 }
+            };
+
+            var result = (await Api.DiffForTestResult(buildId)).EnsureValidResponse();
+            result.Result.Nodes
+                .Where(n => ScreenshotIds.Contains(n.Id))
+                .Aggregate(dict, (counts, node) =>
+                {
+                    counts[node.Status] += 1;
+                    return counts;
+                });
+
+            if (dict[DiffStatus.Queued] > 0)
+            {
+                throw new VisualClientException("Some diffs are not ready");
+            }
+
+            return dict;
         }
     }
 }
